+++ projects/pip-webui2-themes/src/lib/shared/Theme.ts
@@ -1,0 +1,6 @@
+export class Theme {
+    name: string;
+    palette = 'light';
+}
+
+export type Themes = Theme[];