+++ projects/pip-webui2-themes/src/lib/shared/index.ts
@@ -1,0 +1,2 @@
+export * from './Theme';
+export { PipThemesService, defaultTheme } from './themes.service';