+++ projects/pip-webui2-themes/src/lib/themes.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+
+import { PipThemesService } from './shared/themes.service';
+
+@NgModule({
+  imports: [],
+  declarations: [],
+  providers: [PipThemesService]
+})
+export class PipThemesModule { }