+++ projects/pip-webui2-themes/src/public_api.ts
@@ -1,0 +1,6 @@
+/*
+ * Public API Surface of pip-webui2-themes
+ */
+
+export * from './lib/shared';
+export { PipThemesModule } from './lib/themes.module';