+++ src/app/app.module.ts
@@ -1,0 +1,34 @@
+import { BrowserModule } from '@angular/platform-browser';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { NgModule } from '@angular/core';
+import { FlexLayoutModule } from '@angular/flex-layout';
+import { FormsModule } from '@angular/forms';
+import { MatIconModule, MatButtonModule, MatMenuModule, MatToolbarModule, MatSelectModule, MatSidenavModule } from '@angular/material';
+import { PipThemesModule } from 'pip-webui2-themes';
+
+import { AppComponent } from './app.component';
+import { TestModule } from './test/test.module';
+
+@NgModule({
+  declarations: [
+    AppComponent
+  ],
+  imports: [
+    BrowserModule,
+    BrowserAnimationsModule,
+    FlexLayoutModule,
+    FormsModule,
+    MatIconModule,
+    MatButtonModule,
+    MatMenuModule,
+    MatToolbarModule,
+    MatSelectModule,
+    MatSidenavModule,
+
+    TestModule,
+    PipThemesModule
+  ],
+  providers: [],
+  bootstrap: [AppComponent]
+})
+export class AppModule { }