+++ src/index.ts
@@ -1,4 +1,0 @@
-
-export * from './shared/themes.service';
-export * from './shared/themes.model';
-export * from './themes.module';